更新底层依赖版本为2.0；

+++ core/core-form/src/main/java/qsos/core/form/view/adapter/FormFileAdapter.kt
@@ -10,6 +10,7 @@ import qsos.core.form.utils.FormConfigHelper
 import qsos.core.form.view.hodler.FormItemFileItemHolder
 import qsos.lib.base.base.adapter.BaseAdapter
 import qsos.lib.base.base.holder.BaseHolder
+import qsos.lib.base.callback.OnItemListener
 import qsos.lib.base.utils.ToastUtils
 import kotlin.coroutines.CoroutineContext
 
@@ -20,35 +21,35 @@ import kotlin.coroutines.CoroutineContext
 class FormFileAdapter(
         files: ArrayList<Value>,
         private val mJob: CoroutineContext
-) : BaseAdapter<Value>(files) {
+) : BaseAdapter<Value>(files), OnItemListener<Any?> {
 
     override fun getHolder(view: View, viewType: Int): BaseHolder<Value> = FormItemFileItemHolder(view, this)
 
     override fun getLayoutId(viewType: Int): Int = R.layout.form_item_file_item
 
-    override fun onItemClick(view: View, position: Int, obj: Any?) {
-        when (view.id) {
-            R.id.iv_item_form_file_icon -> {
-                /**预览*/
-                FormConfigHelper.previewFile(mContext!!, position, data.map { it.file!! })
-            }
-            R.id.iv_item_form_file_delete -> {
-                /**删除*/
-                if (!data[position].limitEdit) {
-                    CoroutineScope(mJob).dbComplete {
-                        db = { FormDatabase.getInstance().formItemValueDao.delete(data[position]) }
-                        onSuccess = {
-                            data.removeAt(position)
-                            notifyDataSetChanged()
-                        }
-                        onFail = {
-                            ToastUtils.showToastLong(mContext, "删除失败 ${it.message}")
+    override fun onClick(view: View, position: Int, obj: Any?, long: Boolean) {
+        if (!long) {
+            when (view.id) {
+                R.id.iv_item_form_file_icon -> {
+                    /**预览*/
+                    FormConfigHelper.previewFile(mContext!!, position, data.map { it.file!! })
+                }
+                R.id.iv_item_form_file_delete -> {
+                    /**删除*/
+                    if (!data[position].limitEdit) {
+                        CoroutineScope(mJob).dbComplete {
+                            db = { FormDatabase.getInstance().formItemValueDao.delete(data[position]) }
+                            onSuccess = {
+                                data.removeAt(position)
+                                notifyDataSetChanged()
+                            }
+                            onFail = {
+                                ToastUtils.showToastLong(mContext, "删除失败 ${it.message}")
+                            }
                         }
                     }
                 }
             }
         }
     }
-
-    override fun onItemLongClick(view: View, position: Int, obj: Any?) {}
 }
